requirements.txt
Flask>=2.3.0
qrcode[pil]
Flask-Login
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
